fix: message content padding
