+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/messages/item/MessageStatusIndicator.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
-package com.wire.android.ui.home.conversations
+package com.wire.android.ui.home.conversations.messages.item
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.material3.Icon
